Moved galaxy token method
